+++ src/test/ns3tcp/ns3tcp-loss-test-suite.cc
@@ -35,7 +35,7 @@
 #include "ns3/string.h"
 #include "ns3/tcp-header.h"
 #include "ns3/tcp-socket-factory.h"
-#include "ns3/tcp-westwood.h"
+#include "ns3/tcp-westwood-plus.h"
 #include "ns3/test.h"
 #include "ns3/uinteger.h"
 
@@ -146,7 +146,7 @@ Ns3TcpLossTestCase::Ns3TcpLossTestCase()
       m_writeResults(WRITE_PCAP),
       m_writeLogging(WRITE_LOGGING),
       m_needToClose(true),
-      m_tcpModel("ns3::TcpWestwood")
+      m_tcpModel("ns3::TcpWestwoodPlus")
 {
 }
 
@@ -351,15 +351,7 @@ Ns3TcpLossTestCase::DoRun()
 
     std::ostringstream tcpModel;
     tcpModel << "ns3::Tcp" << m_tcpModel;
-    if (m_tcpModel == "WestwoodPlus")
-    {
-        Config::SetDefault("ns3::TcpL4Protocol::SocketType", TypeIdValue(TcpWestwood::GetTypeId()));
-        Config::SetDefault("ns3::TcpWestwood::ProtocolType", EnumValue(TcpWestwood::WESTWOODPLUS));
-    }
-    else
-    {
-        Config::SetDefault("ns3::TcpL4Protocol::SocketType", StringValue(tcpModel.str()));
-    }
+    Config::SetDefault("ns3::TcpL4Protocol::SocketType", StringValue(tcpModel.str()));
 
     Config::SetDefault("ns3::TcpSocket::SegmentSize", UintegerValue(1000));
     Config::SetDefault("ns3::TcpSocket::DelAckCount", UintegerValue(1));
@@ -370,7 +362,7 @@ Ns3TcpLossTestCase::DoRun()
         LogComponentEnableAll(LOG_PREFIX_FUNC);
         LogComponentEnable("Ns3TcpLossTest", LOG_LEVEL_ALL);
         LogComponentEnable("ErrorModel", LOG_LEVEL_DEBUG);
-        LogComponentEnable("TcpWestwood", LOG_LEVEL_ALL);
+        LogComponentEnable("TcpWestwoodPlus", LOG_LEVEL_ALL);
         LogComponentEnable("TcpCongestionOps", LOG_LEVEL_INFO);
         LogComponentEnable("TcpSocketBase", LOG_LEVEL_INFO);
     }
@@ -529,12 +521,6 @@ Ns3TcpLossTestSuite::Ns3TcpLossTestSuite()
     AddTestCase(new Ns3TcpLossTestCase("NewReno", 2), TestCase::QUICK);
     AddTestCase(new Ns3TcpLossTestCase("NewReno", 3), TestCase::QUICK);
     AddTestCase(new Ns3TcpLossTestCase("NewReno", 4), TestCase::QUICK);
-
-    AddTestCase(new Ns3TcpLossTestCase("Westwood", 0), TestCase::QUICK);
-    AddTestCase(new Ns3TcpLossTestCase("Westwood", 1), TestCase::QUICK);
-    AddTestCase(new Ns3TcpLossTestCase("Westwood", 2), TestCase::QUICK);
-    AddTestCase(new Ns3TcpLossTestCase("Westwood", 3), TestCase::QUICK);
-    AddTestCase(new Ns3TcpLossTestCase("Westwood", 4), TestCase::QUICK);
 
     AddTestCase(new Ns3TcpLossTestCase("WestwoodPlus", 0), TestCase::QUICK);
     AddTestCase(new Ns3TcpLossTestCase("WestwoodPlus", 1), TestCase::QUICK);